api/requirements.txt
Flask==3.0.3
mysql-connector-python==8.3.0
bcrypt==4.1.2
Flask-WTF==1.2.1
Flask-Talisman==1.1.0
python-dotenv==1.0.1
gunicorn==21.2.0
Flask-Cors==4.0.0
